feat(website): extract landing page into Svelte 5 components
Extract monolithic 558-line +page.svelte into 11 reusable components:

Reusable:
- TerminalBlock: shared terminal chrome (header + dots + body + copy)
- FeatureCard: glass card with icon, title, accent, description
- StepCard: numbered step card for How It Works
- TimelineStep: numbered timeline item with optional terminal snippet

Sections:
- HeroSection: badge, gradient headline, install terminal, CTAs
- KpiStats: 3 glass metric cards
- LiveDemo: terminal with typing animation (self-contained state)
- HowItWorks: 3 steps with connectors
- FeatureGrid: 9 feature cards in responsive grid
- ComparisonTable: vs CodeRabbit/Copilot/SonarQube
- QuickStart: 4 timeline steps + CTA + footer

- +page.svelte reduced to 47 lines (imports + composition)
- Replaced inline SVGs with @lucide/svelte icons (ArrowRight, Star, etc.)
- GitHub logo kept as inline SVG (trademark)
- All components use Svelte 5 runes ($state, $props, $effect)
- Animation logic preserved (IntersectionObserver for scroll-reveal)

Also adds @lucide/svelte dependency.

Closes #169
