+++ website/src/lib/components/landing/ComparisonTable.svelte
@@ -1,0 +1,83 @@
+<!-- ====== S6 — COMPARISON TABLE ====== -->
+<section class="section section-compact">
+	<h2 class="text-center scroll-reveal text-2xl md:text-3xl font-bold text-[var(--foreground)] -tracking-tight leading-tight">
+		Why developers choose cora
+	</h2>
+	<p class="text-center mt-4 scroll-reveal text-sm text-[var(--muted-foreground)]">
+		Side-by-side with popular code review tools.
+	</p>
+
+	<div class="glass-card scroll-reveal mt-10 p-0 max-w-[56rem] mx-auto overflow-hidden">
+		<div class="overflow-x-auto">
+			<table class="compare-table">
+				<thead>
+					<tr>
+						<th>Feature</th>
+						<th class="cora-col">cora</th>
+						<th>CodeRabbit</th>
+						<th>Copilot</th>
+						<th>SonarQube</th>
+					</tr>
+				</thead>
+				<tbody>
+					<tr>
+						<td>BYOK</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td class="text-[var(--muted-foreground)]">&mdash;</td>
+					</tr>
+					<tr>
+						<td>Self-hosted</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-check">&#10003;</span></td>
+					</tr>
+					<tr>
+						<td>Gitea</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-check">&#10003;</span></td>
+					</tr>
+					<tr>
+						<td>CLI</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+					</tr>
+					<tr>
+						<td>Pre-commit</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+						<td><span class="symbol-cross">&#10007;</span></td>
+					</tr>
+					<tr>
+						<td>SARIF</td>
+						<td class="cora-col"><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-check">&#10003;</span></td>
+						<td><span class="symbol-check">&#10003;</span></td>
+					</tr>
+					<tr>
+						<td class="font-semibold">Cost</td>
+						<td class="cora-col font-semibold">Free + API</td>
+						<td>$12-39/mo</td>
+						<td>$10-39/mo</td>
+						<td>Free / $150+</td>
+					</tr>
+					<tr>
+						<td>License</td>
+						<td class="cora-col">MIT</td>
+						<td>Apache 2.0</td>
+						<td>Proprietary</td>
+						<td>LGPL</td>
+					</tr>
+				</tbody>
+			</table>
+		</div>
+	</div>
+</section>
