+++ website/src/lib/components/landing/KpiStats.svelte
@@ -1,0 +1,20 @@
+<!-- ====== S2 — KPI STATS ====== -->
+<section class="section section-compact">
+	<p class="text-center mb-10 scroll-reveal text-xs font-medium text-[var(--muted-foreground)] uppercase tracking-widest">
+		Trusted by developers who ship fast
+	</p>
+	<div class="grid grid-cols-1 md:grid-cols-3 gap-4 max-w-3xl mx-auto">
+		<div class="glass-card text-center scroll-reveal py-8">
+			<div class="text-2xl md:text-3xl font-bold text-[var(--foreground)] -tracking-tight leading-tight">5</div>
+			<div class="text-sm text-[var(--muted-foreground)] mt-2">AI Providers</div>
+		</div>
+		<div class="glass-card text-center scroll-reveal py-8 [transition-delay:100ms]">
+			<div class="text-2xl md:text-3xl font-bold text-[var(--foreground)] -tracking-tight leading-tight">&lt; 3s</div>
+			<div class="text-sm text-[var(--muted-foreground)] mt-2">Review Time</div>
+		</div>
+		<div class="glass-card text-center scroll-reveal py-8 [transition-delay:200ms]">
+			<div class="text-2xl md:text-3xl font-bold text-[var(--foreground)] -tracking-tight leading-tight">Zero</div>
+			<div class="text-sm text-[var(--muted-foreground)] mt-2">Config Required</div>
+		</div>
+	</div>
+</section>
